Eliminated map View Option
