+++ app/Valapas/.idea/dictionaries/dayna.xml
@@ -1,0 +1,7 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="dayna">
+    <words>
+      <w>valapas</w>
+    </words>
+  </dictionary>
+</component>
