+++ app/Valapas/app/src/main/java/com/group3/valapas/UserFragments/UserBookingsMostRecent.java
@@ -24,7 +24,6 @@ public class UserBookingsMostRecent extends Fragment
     private Button btnFavoritesMostRecent;
     private Button btnFavoritesMostUsed;
     private Button btnFavoritesPrice;
-    private Button btnOpenMap;
     private Button btnProfilePage;
 
     @Nullable
@@ -45,7 +44,6 @@ public class UserBookingsMostRecent extends Fragment
         btnFavoritesMostRecent = (Button) view.findViewById(R.id.user_favorites_recent);
         btnFavoritesMostUsed = (Button) view.findViewById(R.id.user_favorites_used);
         btnFavoritesPrice = (Button) view.findViewById(R.id.user_favorites_price);
-        btnOpenMap = (Button) view.findViewById(R.id.open_map_btn);
         btnProfilePage = (Button) view.findViewById(R.id.user_profile_page);
 
         btnBookingsPage.setOnClickListener(new View.OnClickListener()
@@ -156,21 +154,12 @@ public class UserBookingsMostRecent extends Fragment
             }
         });
 
-        btnOpenMap.setOnClickListener(new View.OnClickListener()
-        {
-            @Override
-            public void onClick(View view)
-            {
-                ((UserActivity)getActivity()).setViewPager(12);
-            }
-        });
-
         btnProfilePage.setOnClickListener(new View.OnClickListener()
         {
             @Override
             public void onClick(View view)
             {
-                ((UserActivity)getActivity()).setViewPager(13);
+                ((UserActivity)getActivity()).setViewPager(12);
             }
         });
