--- app/Valapas/app/src/main/res/layout/browse_map_view.xml
@@ -1,147 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-
-    android:layout_marginTop="40dp"
-    android:layout_marginLeft="40dp"
-    android:layout_marginRight="40dp"
-    android:layout_marginBottom="40dp"
-
-    android:orientation="vertical"
-    tools:context="BrowseFragments.BrowseMap">
-
-    <LinearLayout
-        android:id="@+id/search_all_header"
-
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-
-        android:orientation="horizontal"
-
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toTopOf="@id/search_options"
-        >
-
-        <EditText
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-
-            android:importantForAutofill="no"
-            android:inputType="text"
-            tools:ignore="LabelFor"
-            />
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:id="@+id/search_options"
-
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-
-        android:orientation="horizontal"
-        android:layout_marginTop="10dp"
-
-        app:layout_constraintTop_toBottomOf="@id/search_all_header"
-        app:layout_constraintBottom_toTopOf="@id/search_all_results"
-        >
-
-        <Button
-            android:id="@+id/user_browse_category"
-            style="@style/Widget.AppCompat.Button.Borderless"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginEnd="2dp"
-            android:layout_marginRight="2dp"
-            android:layout_weight="1"
-            android:onClick="selectCategory"
-            android:text="@string/category"
-            android:textSize="10sp" />
-
-        <Button
-            android:id="@+id/user_browse_price"
-            style="@style/Widget.AppCompat.Button.Borderless"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_marginRight="1dp"
-            android:layout_marginLeft="1dp"
-            android:layout_weight="1"
-            android:onClick="selectPrice"
-            android:text="@string/price"
-            android:textSize="10sp" />
-
-        <Button
-            android:id="@+id/user_browse_popular"
-            style="@style/Widget.AppCompat.Button.Borderless"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center_horizontal"
-            android:layout_marginStart="2dp"
-            android:layout_marginLeft="2dp"
-            android:layout_weight="1"
-            android:onClick="selectPopular"
-            android:text="@string/popular"
-            android:textSize="10sp" />
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:id="@+id/search_all_results"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="1"
-
-        android:layout_marginTop="10dp"
-
-        app:layout_constraintTop_toBottomOf="@+id/search_options"
-        app:layout_constraintBottom_toTopOf="@id/browse_page_footer">
-
-        <ListView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:verticalScrollbarPosition="left"/>
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:id="@+id/browse_page_footer"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="10dp"
-
-        android:orientation="horizontal"
-
-        app:layout_constraintTop_toBottomOf="@+id/search_all_results"
-        app:layout_constraintBottom_toBottomOf="parent">
-
-        <Button
-            android:id="@+id/browse_page_btn"
-            style="@style/Widget.AppCompat.Button.Borderless"
-            android:layout_width="0dp"
-            android:layout_weight="3"
-            android:layout_height="wrap_content"
-            android:layout_marginEnd="2dp"
-            android:layout_marginRight="2dp"
-            android:text="@string/browse_page"
-            android:textSize="9sp"
-            />
-
-        <Button
-            android:id="@+id/open_map_btn"
-            style="@style/Widget.AppCompat.Button.Borderless"
-            android:layout_width="0dp"
-            android:layout_weight="1"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="2dp"
-            android:layout_marginLeft="2dp"
-            android:text="@string/open_map"
-            android:textSize="9sp"
-            />
-
-    </LinearLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
