+++ app/Valapas/app/src/main/res/layout/browse_profile.xml
@@ -1,12 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/footer_layout"
     android:layout_width="match_parent"
-    android:layout_height="75dp"
-    android:layout_gravity="bottom"
-    android:orientation="vertical">
+    android:layout_height="match_parent"
+    xmlns:tools="http://schemas.android.com/tools"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    tools:context="BrowseFragments.BrowseProfile">
 
     <LinearLayout
         android:id="@+id/footer"
@@ -42,13 +40,13 @@
             android:layout_height="wrap_content"
             android:layout_weight="1"
             android:background="@color/DarkGray"
-            android:drawableTop="@drawable/my_profile_icon_white"
+            android:drawableTop="@drawable/my_profile_icon_gold"
             android:paddingTop="15dp"
             android:paddingBottom="15dp"
             android:text="@string/user_profile"
             android:textAlignment="center"
             android:textAllCaps="true"
-            android:textColor="@color/White"
+            android:textColor="@color/SelectedGold"
             android:textSize="9sp"
             android:textStyle="normal"
             android:visibility="visible"
