--- app/Valapas/app/src/main/res/layout/user_open_map_view.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".UserFragments.UserOpenMap">
-
-</androidx.constraintlayout.widget.ConstraintLayout>
